+++ axum-core/src/extract/mod.rs
@@ -36,7 +36,7 @@ mod private {
 ///
 /// See [`axum::extract`] for more general docs about extraxtors.
 ///
-/// [`axum::extract`]: https://docs.rs/axum/0.6.0-rc.1/axum/extract/index.html
+/// [`axum::extract`]: https://docs.rs/axum/0.6.0-rc.2/axum/extract/index.html
 #[async_trait]
 pub trait FromRequestParts<S>: Sized {
     /// If the extractor fails it'll use this "rejection" type. A rejection is
@@ -99,7 +99,7 @@ pub trait FromRequestParts<S>: Sized {
 /// This ensures your extractor is as flexible as possible.
 ///
 /// [`http::Request<B>`]: http::Request
-/// [`axum::extract`]: https://docs.rs/axum/0.6.0-rc.1/axum/extract/index.html
+/// [`axum::extract`]: https://docs.rs/axum/0.6.0-rc.2/axum/extract/index.html
 #[async_trait]
 pub trait FromRequest<S, B, M = private::ViaRequest>: Sized {
     /// If the extractor fails it'll use this "rejection" type. A rejection is