+++ axum/CHANGELOG.md
@@ -17,9 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **added**: Add `middleware::from_fn_with_state` and
   `middleware::from_fn_with_state_arc` to enable running extractors that require
-  state
+  state ([#1342])
 
 [#1327]: https://github.com/tokio-rs/axum/pull/1327
+[#1342]: https://github.com/tokio-rs/axum/pull/1342
 
 # 0.6.0-rc.1 (23. August, 2022)